dashboard/backend/requirements.txt
Django==1.11.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-auth==0.9.3
django-allauth==0.35.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
elasticsearch==5.5.1
elasticsearch-dsl==5.3.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
urllib3==1.21.1
